prevent spinning down of parallel cluster on run_parallel when desired

+++ vowpal_platypus/vw.py
@@ -465,7 +465,7 @@ def run(model, filename=None, train_filename=None, predict_filename=None, line_f
                     header=header)
 
 
-def run_parallel(vw_models, core_fn):
+def run_parallel(vw_models, core_fn, spindown=True):
     num_cores = len(vw_models) if isinstance(vw_models, collections.Sequence) else 1
     if num_cores > 1:
         os.system("spanning_tree")
@@ -480,8 +480,9 @@ def run_parallel(vw_models, core_fn):
                 raise e
         pool = Pool(num_cores)
         results = pool.map(run_fn, vw_models) # TODO: Integrate into `run`
-        os.system('killall vw')
-        os.system('killall spanning_tree')
+        if spindown:
+            os.system('killall vw')
+            os.system('killall spanning_tree')
         return results
     else:
         return core_fn(vw_models[0] if isinstance(vw_models, collections.Sequence) else vw_models)
